// src/app/shared/models/developper.ts
export interface Developer {
  id: number;
  user_id: number;
  Nomdev: string;
  Prenomdev: string;
  CompetencesTechniques: string;
  Experience?: number;
  Niveau?: string;
  Disponibilite?: string;
  Ville?: string;
  Pays?: string;
  Github?: string;
  TarifJournalier?: number;
  Portfolio?: string;
  moyenneNote?: number;
  Pseudo: string;
  Emaildev: string;
  Telephone: string;
  DateNaissance: string;
  Biographie: string;
}